update: move parse_yaml from cli to api
